app-nginx - 0.27 - nginx-go: verify config - 13

+++ dockrepo/sysdata/nginx/nginxconfig.sh
@@ -13,6 +13,8 @@ rsa_key_size=4096
 data_path="${ConfigWorkDir}data/certbot"
 staging=0 # Set to 1 if you are testing your setup to avoid hitting request limits
 #
+DebugMode=1
+#
 #  Funzioni utili
 ###############################################################################
 avanti(){
@@ -30,7 +32,37 @@ pause() {
 }
 #
 ##########################################################################
+bugmessage(){
+# stampa di un avviso di errore, se DebugMode -gt 0
 #
+#  $1 :  ERROR oppure WARNING oppure NOTE
+#  $2 :  codice di errore
+#  stampa piccola storia  <-- vecchio debuglog !!!                              DA RIVEDERE
+#  $3 :  avviso all'utente, come '... il file xxx e' essenziale ...'
+#
+  if [ ${DebugMode} -gt 0 ]; then
+    echo 
+    echo "------------------------------------------------------"
+    echo -e "${Red} $1 $2 - debug message: ${Std}"
+    echo -e "$3"
+    echo "------------------------------------------------------"
+    echo 
+  fi
+}
+#
+##########################################################################
+bugprint(){
+# stampa di un avviso, ma solo se DebugMode -gt 0
+#
+#  $1 :  avviso all'utente, come: 'ho fatto questo o quello'
+#
+  if [ ${DebugMode} -gt 0 ]; then
+    echo $1
+  fi
+}
+#
+##########################################################################
+#                                                                     MAIN
 echo "--------------------------------------------------------------"
 echo "SCRIPT: $0"
 echo
@@ -61,7 +93,7 @@ pause
 if [ -d "$data_path" ]; then
   read -p "Existing data found for $domains. Continue and replace existing certificate? (y/N) " decision
   if [ "$decision" != "Y" ] && [ "$decision" != "y" ]; then
-    exit
+    exit 0
   fi
 fi
 
@@ -73,29 +105,54 @@ if [ ! -e "$data_path/conf/options-ssl-nginx.conf" ] || [ ! -e "$data_path/conf/
   echo "... Downloading recommended TLS parameters ..."
   mkdir -p "$data_path/conf"
   curl -s https://raw.githubusercontent.com/certbot/certbot/master/certbot-nginx/certbot_nginx/_internal/tls_configs/options-ssl-nginx.conf > "$data_path/conf/options-ssl-nginx.conf"
+  RetValue=$?
+  if [ ${RetValue} -gt 0 ]; then
+    bugmessage ERROR "${RetValue}" "    config.sh: get ssl params"
+    exit 31 #                                              ssl config error  -->  ERRORE 31
+  fi
   curl -s https://raw.githubusercontent.com/certbot/certbot/master/certbot/certbot/ssl-dhparams.pem > "$data_path/conf/ssl-dhparams.pem"
+  RetValue=$?
+  if [ ${RetValue} -gt 0 ]; then
+    bugmessage ERROR "${RetValue}" "    config.sh: get ssl params"
+    exit 32 #                                              ssl config error  -->  ERRORE 32
+  fi
   echo
 fi
 
 echo "### Creating dummy certificate for $domains ..."
 path="/etc/letsencrypt/live/$domains"
 mkdir -p "$data_path/conf/live/$domains"
-docker-compose -f nginx.yml run --rm --entrypoint "\
+docker-compose -f nginx-certbot.yml run --rm --entrypoint "\
   openssl req -x509 -nodes -newkey rsa:$rsa_key_size -days 1\
     -keyout '$path/privkey.pem' \
     -out '$path/fullchain.pem' \
     -subj '/CN=localhost'" certbot
+RetValue=$?
+if [ ${RetValue} -gt 0 ]; then
+  bugmessage ERROR "${RetValue}" "    config.sh: get ssl params"
+  exit 33 #                                                ssl get RSA keys  -->  ERRORE 33
+fi
 
 echo
 echo "### Starting nginx ..."
-docker-compose -f nginx.yml up --force-recreate -d nginx
+docker-compose -f nginx-certbot.yml up --force-recreate -d nginx
+RetValue=$?
+if [ ${RetValue} -gt 0 ]; then
+  bugmessage ERROR "${RetValue}" "    config.sh: get ssl params"
+  exit 34 #                                                   Starting nginx -->  ERRORE 34
+fi
 echo
 
 echo "### Deleting dummy certificate for $domains ..."
-docker-compose -f nginx.yml run --rm --entrypoint "\
+docker-compose -f nginx-certbot.yml run --rm --entrypoint "\
   rm -Rf /etc/letsencrypt/live/$domains && \
   rm -Rf /etc/letsencrypt/archive/$domains && \
   rm -Rf /etc/letsencrypt/renewal/$domains.conf" certbot
+RetValue=$?
+if [ ${RetValue} -gt 0 ]; then
+  bugmessage ERROR "${RetValue}" "    config.sh: delete certificates"
+  exit 35 #                                                   Starting nginx -->  ERRORE 35
+fi
 echo
 
 echo "### Requesting certificate for $domains ..."
@@ -114,7 +171,7 @@ esac
 # Enable staging mode if needed
 if [ $staging != "0" ]; then staging_arg="--staging"; fi
 
-docker-compose -f nginx.yml run --rm --entrypoint "\
+docker-compose -f nginx-certbot.yml run --rm --entrypoint "\
   certbot certonly --webroot -w /var/www/certbot \
     $staging_arg \
     $email_arg \
@@ -122,15 +179,34 @@ docker-compose -f nginx.yml run --rm --entrypoint "\
     --rsa-key-size $rsa_key_size \
     --agree-tos \
     --force-renewal" certbot
+RetValue=$?
+if [ ${RetValue} -gt 0 ]; then
+  bugmessage ERROR "${RetValue}" "    config.sh: run nginx & certbot"
+  exit 35 #                                              run nginx & certbot -->  ERRORE 35
+fi
 echo
 
 echo "... Reloading nginx ..."
-docker-compose -f nginx.yml exec nginx nginx -s reload
+docker-compose -f nginx-certbot.yml exec nginx nginx -s reload
+RetValue=$?
+if [ ${RetValue} -gt 0 ]; then
+  bugmessage ERROR "${RetValue}" "    config.sh: reload nginx"
+  exit 36 #                                              run nginx & certbot -->  ERRORE 36
+fi
+echo "NGINX proxy is started!"
+
+echo "... Reset nginx & certbot"
+docker-compose -f nginx-certbot.yml down
+RetValue=$?
+if [ ${RetValue} -gt 0 ]; then
+  bugmessage ERROR "${RetValue}" "    nginx & certbot down"
+  exit 37 #                                             nginx & certbot down -->  ERRORE 37
+fi
+echo "NGINX proxy and certbot now DOWN"
+
 echo
 echo "================================================================"
-echo "NGINX proxy is started!"
-echo
-echo "Please verify nginx is OK"
+echo "Config script OK"
 echo
 pause
-return 0
+exit 0
